third_party/blink/web_tests/svg/dynamic-updates/SVGFEMergeNodeElement-dom-in-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in' attribute of the SVGFEMergeNodeElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS mergeNode1.getAttribute('in') is "SourceGraphic"
PASS mergeNode1.getAttribute('in') is "off"
PASS successfullyParsed is true

TEST COMPLETE
